# apps/backend/docs/swagger.yaml
basePath: /
definitions:
...
  internal_folder.FolderResult:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      parent_folder_id:
        type: string
      starred:
        type: boolean
      updated_at:
        type: string
    type: object
  internal_folder.createRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
      parent_folder_id:
        type: string
    required:
    - name
    type: object
  internal_folder.moveRequest:
    properties:
      parent_folder_id:
        type: string
    type: object
  internal_folder.renameRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
  internal_folder.starredRequest:
    properties:
      starred:
        type: boolean
    type: object
host: localhost:4000
info:
  contact: {}
  description: CriByte — self-hosted Google Drive clone
  title: CriByte API
  version: "1.0"
paths:
...
  /api/folders:
    post:
      consumes:
      - application/json
      description: Create a new folder. Auto-resolves name conflicts.
      parameters:
      - description: Folder name and optional parent
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_folder.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_folder.FolderResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Create a folder
      tags:
      - folders
  /api/folders/{id}:
    get:
      description: Get metadata for a folder by its ID
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_folder.FolderResult'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Get folder metadata
      tags:
      - folders
    patch:
      consumes:
      - application/json
      description: Rename a folder. Auto-resolves name conflicts.
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: New name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_folder.renameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_folder.FolderResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Rename a folder
      tags:
      - folders
  /api/folders/{id}/contents:
    get:
      description: List subfolders inside a folder
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_folder.FolderResult'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: List folder contents
      tags:
      - folders
  /api/folders/{id}/move:
    post:
      consumes:
      - application/json
      description: Move a folder to a different parent
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: Target parent folder
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_folder.moveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_folder.FolderResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Move a folder
      tags:
      - folders
  /api/folders/{id}/permanent-delete:
    post:
      description: Delete a folder permanently from trash. Folder must be in trash
        first.
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Permanently delete a folder
      tags:
      - folders
  /api/folders/{id}/restore:
    post:
      description: Restore a trashed folder to its original location
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_folder.FolderResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Restore a folder from trash
      tags:
      - folders
  /api/folders/{id}/star:
    post:
      consumes:
      - application/json
      description: Toggle the starred/favorite status of a folder
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: Starred status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_folder.starredRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_folder.FolderResult'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Star or unstar a folder
      tags:
      - folders
  /api/folders/{id}/trash:
    delete:
      description: Soft-delete a folder and all its descendants
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_folder.FolderResult'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Trash a folder
      tags:
      - folders
  /api/folders/starred:
    get:
      description: List all starred folders for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_folder.FolderResult'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: List starred folders
      tags:
      - folders
  /api/folders/trash:
    get:
      description: List all folders in trash for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_folder.FolderResult'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: List trashed folders
      tags:
      - folders
schemes:
- http
- https
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"